Partially revive WebRTC functionality in Swell
The WebRTC functionality does not work as-is, and
this commit attempts to revive the STUN/TURN
server checks in the app.

- Rewrite the logic to perform STUN/TURN server
tests, and automatically close the connection once
the check finishes
- Display the results in the response pane
- Remove the need to send response within the
request details container

NOTE: The frontend currently does not work,
because there is no infrastructure to properly
sanitize and parse an array of objects. If this
functionality is being worked on in the future,
developers may want to consider using input
texts rather than codemirror for easier
formatting in the backend.