Bug fix: Removing GraphQL from workspace crashes app
The way that the `reqRes` is setup in `GraphQL` creates
an issue where in a query and mutation scenario, the
`protocol` is incorrectly hardcoded as `ws://` (i.e. WebSocket)
even though the URL starts with `http://`.

This commit rewrites the logic in `GraphQLComposer.tsx`
to remove unnecessary redundancies and update
`closeReqRes` to ensure GraphQL subscriptions are
closed properly