Update GraphQL Server to using apollo server
`apollo-server-express` is scheduled to be deprecated
at the end of 2023. This commit updates the GraphQL
server to use `@apollo/server` instead.

Note that Apollo server 4 requires top-level async/await
for the server to start properly, which can only be done
in a module. Since most backend logic in Swell uses
CommonJS, we need to change this file to `.mjs` to make
things work. The downside to making this as a `.mjs` file
is that it would not be automatically compatible with the
rest of the backend code that uses CommonJS.
Since this is only for testing and as a standalone server,
we should not need to configure the bundler to accept
both. If we do in the future, we may need to research
on how to do so in webpack.