Fix bugs associated with disabling nav bar buttons
The current path when the app first started is the
HTTP/2 page. The local `state`, however, does not
point to the same thing.

This commit ensures the initial `curPage` state in
`ProtocolSelect.tsx` is consistent with `currPath`
and the request fields and contents are updated
accordingly.

Tests are also updated to handle the disabling
feature